requirements.txt
numpy
pandas
#tensorflow==2.10.0
fastapi==0.109.0
uvicorn
typing_extensions==4.8.0
pytz
pillow
python-multipart
google-cloud-bigquery
